+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-10-01
+
+### Added
+
+- Pipeline for interaction tags classifier (TagsClassifierPipeline).
+
 ## 2020-09-24
 
 ### Added